Update file store timeout value
After upgrading from rails 3 to rails 4 we hit timeout error with the file store lock. We're compiling assets on a 32core machine, so that doesn't give much time given the timeout is only 1000ms. Raising the timeout has fixed this problem for us and doesn't seem to have any ill effects.